Added repository for fetching initial settings

+++ BoardOrder.Domain/BoardOrderItem.cs
@@ -3,7 +3,7 @@
 		public BoardOrderItem(string name)
 			: this(name, 0, 0) { }
 
-		public BoardOrderItem(string name, int costModifier, int workdaysModifier) {
+		public BoardOrderItem(string name, double costModifier, double workdaysModifier) {
 			this.Name = Name;
 			this.CostModifier = CostModifier;
 			this.WorkdaysModifier = WorkdaysModifier;
